update user group

+++ app/Http/Controllers/RehabilitasiController.php
@@ -14,6 +14,7 @@ use Str;
 use Ramsey\Uuid\Uuid;
 use App\Rules\CheckTahunRehab;
 use Image;
+use Auth;
 
 class RehabilitasiController extends Controller
 {
@@ -253,10 +254,15 @@ class RehabilitasiController extends Controller
                          return $data;
                      })
                     ->editColumn('aksi', function ($row) {
-                        $data = '
-                              <a title="Ubah Data" class="btn btn-success btn-sm" onclick="ubah_rehabilitasi(\''.$row->id_rehabilitasi.'\')"> <i class="fas fa-edit text-white"></i></a>
-                              <a title="Hapus Data" class="btn btn-danger btn-sm" onclick="hapus_rehabilitasi(\''.$row->id.'\')"> <i class="fas fa-trash-alt text-white"></i></a>
-                         ';
+
+                         if (Auth::user()->group != 2) {
+                             $data = '
+                                   <a title="Ubah Data" class="btn btn-success btn-sm" onclick="ubah_rehabilitasi(\''.$row->id_rehabilitasi.'\')"> <i class="fas fa-edit text-white"></i></a>
+                                   <a title="Hapus Data" class="btn btn-danger btn-sm" onclick="hapus_rehabilitasi(\''.$row->id.'\')"> <i class="fas fa-trash-alt text-white"></i></a>
+                              ';
+                         }else{
+                              $data = '';
+                         }
 
                         return $data;
                     })
